pyproject.toml
[project]
name = "anthropic-autogen"
version = "0.1.0"
description = "A Python framework for building autonomous agent systems powered by Anthropic's Claude models"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    # Core Dependencies
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.2",
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "websockets>=12.0",
    "click>=8.1.7",
    "pyyaml>=6.0.2",
    "numpy>=1.24.0",
    "pandas>=2.1.3",
    "matplotlib>=3.8.2",
    "openai>=1.3.5",
    "scikit-learn>=1.3.2",  # Using scikit-learn instead of faiss for now
    "tenacity>=8.2.3",
    "pylint>=3.0.2",
    "black>=23.11.0",
    "isort>=5.12.0",
    "gitpython>=3.1.40",
    "packaging>=23.2",
    "aiofiles>=23.2.1",  # Async file operations
    "psutil>=5.9.6",  # Process management
    "beautifulsoup4>=4.12.2",  # HTML parsing
    "aiohttp>=3.9.1",  # Async HTTP client
    "playwright>=1.40.0",  # Modern web automation

    # AI/LLM Dependencies
    "anthropic>=0.3.11",
    "autogen-core>=0.4.0.dev6",
    "autogen-ext>=0.4.0.dev6",
    "grpcio==1.62.0",  # Pin to version compatible with Python 3.12

    # Data Analysis Dependencies
    "seaborn>=0.13.0",
    "scipy>=1.11.3",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "hypothesis>=6.92.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=anthropic_autogen --cov-report=term-missing"

[project.scripts]
cli = "anthropic_autogen.cli:cli"
